✅ [chore] #54 파일 분리
